Revert "[DF-5671] Used 64bits for components mask"
This reverts commit 9514787a41c47158e76ba711f7eabb143ecb5e13 [formerly 97e95921f7ec57a0d1d46fb9230e6144ff2e3c7e].


Former-commit-id: 087fe01c5614873836e7f0de2273367df3636b45

+++ Sources/Internal/Entity/SceneSystem.h
@@ -46,8 +46,8 @@ public:
     SceneSystem(Scene * scene);
     virtual ~SceneSystem();
     
-    inline void SetRequiredComponents(uint64 requiredComponents);
-    inline uint64 GetRequiredComponents() const;
+    inline void SetRequiredComponents(uint32 requiredComponents);
+    inline uint32 GetRequiredComponents() const;
     
     /**
         \brief  This function is called when any entity registered to scene.
@@ -131,7 +131,7 @@ protected:
 	inline Scene * GetScene() const;
 
 private:
-    uint64 requiredComponents;
+    uint32 requiredComponents;
 	Scene * scene;
 
 	bool locked;
@@ -143,12 +143,12 @@ inline Scene * SceneSystem::GetScene() const
     return scene;
 }
 
-inline void SceneSystem::SetRequiredComponents(uint64 _requiredComponents)
+inline void SceneSystem::SetRequiredComponents(uint32 _requiredComponents)
 {
     requiredComponents = _requiredComponents;
 }
 
-inline uint64 SceneSystem::GetRequiredComponents() const
+inline uint32 SceneSystem::GetRequiredComponents() const
 {
     return requiredComponents;
 }
