+++ Sources/Internal/Scene3D/Entity.cpp
@@ -59,7 +59,7 @@ namespace DAVA
 const int COMPONENT_COUNT_V6 = 18;
 const int COMPONENTS_IN_VECTOR_COUNT = 3;
     
-uint64 vectorComponentsMask = (1 << Component::TRANSFORM_COMPONENT) | (1 << Component::RENDER_COMPONENT) | (1 << Component::LOD_COMPONENT);
+uint32 vectorComponentsMask = (1 << Component::TRANSFORM_COMPONENT) | (1 << Component::RENDER_COMPONENT) | (1 << Component::LOD_COMPONENT);
 
 // Property Names.
 const char* Entity::SCENE_NODE_IS_SOLID_PROPERTY_NAME = "editor.isSolid";
